+++ src/data-workspace/data-workspace.js
@@ -10,7 +10,6 @@ import {
     getDataSetById,
     getCategoryComboById,
 } from './selectors.js'
-import { hashArraysInObject } from './utils.js'
 
 const query = {
     dataSet: {
@@ -39,30 +38,6 @@ const dataValueQuery = {
             orgUnit: orgUnitId,
             attributeOptionCombo: attributeOptionComboId,
         }),
-    },
-}
-
-const metadataQuery = {
-    metadata: {
-        resource: 'metadata',
-        params: {
-            // Note: on dataSet.dataSetElement, the categoryCombo property is
-            // included because it can mean it's overriding the data element's
-            // native categoryCombo. It can sometimes be absent from the data
-            // set element
-            'dataSets:fields':
-                'id,displayFormName,formType,dataSetElements[dataElement,categoryCombo],categoryCombo,sections~pluck',
-            'dataElements:fields': 'id,displayFormName,categoryCombo,valueType',
-            'sections:fields':
-                'id,displayName,sortOrder,showRowTotals,showColumnTotals,disableDataElementAutoGroup,greyedFields[id],categoryCombos~pluck,dataElements~pluck,indicators~pluck',
-            'categoryCombos:fields':
-                'id,skipTotal,categories~pluck,categoryOptionCombos~pluck,isDefault',
-            'categories:fields': 'id,displayFormName,categoryOptions~pluck',
-            'categoryOptions:fields':
-                'id,displayFormName,categoryOptionCombos~pluck,categoryOptionGroups~pluck,isDefault',
-            'categoryOptionCombos:fields':
-                'id,categoryOptions~pluck,categoryCombo,name',
-        },
     },
 }
 
@@ -98,6 +73,7 @@ function mapDataValuesToFormInitialValues(dataValues) {
 }
 
 // TODO: this should probably be handled by useContextSelection-hook
+// should not need this when api support CC and CP instead of cocId
 const useAttributeOptionCombo = () => {
     const { available, metadata } = useMetadata()
     const [{ dataSetId, attributeOptionComboSelection }] = useContextSelection()
@@ -141,14 +117,7 @@ export const DataWorkspace = () => {
         },
     ])
 
-    const { available, metadata, setMetadata } = useMetadata()
-    useQuery([metadataQuery], {
-        onSuccess: (metadata) => {
-            const hashed =
-                metadata?.metadata && hashArraysInObject(metadata.metadata)
-            setMetadata(hashed)
-        },
-    })
+    const { available, metadata } = useMetadata()
 
     const dataValuesFetch = useDataValues()
     console.log({ metadata }, { dataValues: dataValuesFetch.data }, { dataSet: dataSetFetch.data })
