+++ src/data-workspace/metadata-context.js
@@ -1,4 +1,7 @@
+import PropTypes from 'prop-types'
 import React, { useState, useMemo, useContext, useCallback } from 'react'
+import { useQuery } from 'react-query'
+import { hashArraysInObject } from './utils.js'
 
 const cachedMetadata = () => {
     try {
@@ -42,4 +45,46 @@ export const MetadataContextProvider = ({ children }) => {
     )
 }
 
-export const useMetadata = () => useContext(MetadataContext)
+MetadataContextProvider.propTypes = {
+    children: PropTypes.node,
+}
+
+const metadataQuery = {
+    metadata: {
+        resource: 'metadata',
+        params: {
+            // Note: on dataSet.dataSetElement, the categoryCombo property is
+            // included because it can mean it's overriding the data element's
+            // native categoryCombo. It can sometimes be absent from the data
+            // set element
+            'dataSets:fields':
+                'id,displayFormName,formType,dataSetElements[dataElement,categoryCombo],categoryCombo,sections~pluck',
+            'dataElements:fields': 'id,displayFormName,categoryCombo,valueType',
+            'sections:fields':
+                'id,displayName,sortOrder,showRowTotals,showColumnTotals,disableDataElementAutoGroup,greyedFields[id],categoryCombos~pluck,dataElements~pluck,indicators~pluck',
+            'categoryCombos:fields':
+                'id,skipTotal,categories~pluck,categoryOptionCombos~pluck,isDefault',
+            'categories:fields': 'id,displayFormName,categoryOptions~pluck',
+            'categoryOptions:fields':
+                'id,displayFormName,categoryOptionCombos~pluck,categoryOptionGroups~pluck,isDefault',
+            'categoryOptionCombos:fields':
+                'id,categoryOptions~pluck,categoryCombo,name',
+        },
+    },
+}
+
+export const useMetadata = () => {
+    const context = useContext(MetadataContext)
+
+    useQuery([metadataQuery], {
+        onSuccess: (metadata) => {
+            const hashed =
+                metadata?.metadata && hashArraysInObject(metadata.metadata)
+            context.setMetadata(hashed)
+        },
+        staleTime: 60 * 24 * 1000,
+        refetchOnWindowFocus: false,
+    })
+
+    return context
+}
